+++ ehpcinstant-20230701/alibabacloud_ehpcinstant20230701/models/_list_pools_response_body.py
@@ -82,19 +82,22 @@ class ListPoolsResponseBody(DaraModel):
 class ListPoolsResponseBodyPoolList(DaraModel):
     def __init__(
         self,
+        create_time: str = None,
         is_default: bool = None,
-        max_exector_num: int = None,
+        max_executor_num: int = None,
         pool_name: str = None,
         priority: int = None,
+        scheduling_policy_id: str = None,
         status: str = None,
+        update_time: str = None,
     ):
+        self.create_time = create_time
         # Indices whether the resource pool is the default resource pool. Valid values:
         # 
         # *   **true**
         # *   **false**
         self.is_default = is_default
-        # The maximum number of execution nodes that can run concurrently in a resource pool.
-        self.max_exector_num = max_exector_num
+        self.max_executor_num = max_executor_num
         # The name of the resource pool.
         # 
         # *   The value can be up to 15 characters in length.
@@ -105,6 +108,7 @@ class ListPoolsResponseBodyPoolList(DaraModel):
         # *   You can set a priority in the range of 1 to 99. The default value is 1, which is the lowest priority.
         # *   Jobs submitted to a resource pool with a higher priority level value will be scheduled before pending jobs in a resource pool with a lower priority level value, and the priority level of the resource pool takes precedence over the priority of the job.
         self.priority = priority
+        self.scheduling_policy_id = scheduling_policy_id
         # The status of the resource pool. Valid values:
         # 
         # *   Creating: The resource pool is being created.
@@ -113,6 +117,7 @@ class ListPoolsResponseBodyPoolList(DaraModel):
         # *   Working: The resource pool is working.
         # *   Deleted: The resource pool is deleted.
         self.status = status
+        self.update_time = update_time
 
     def validate(self):
         pass
@@ -122,11 +127,14 @@ class ListPoolsResponseBodyPoolList(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.create_time is not None:
+            result['CreateTime'] = self.create_time
+
         if self.is_default is not None:
             result['IsDefault'] = self.is_default
 
-        if self.max_exector_num is not None:
-            result['MaxExectorNum'] = self.max_exector_num
+        if self.max_executor_num is not None:
+            result['MaxExecutorNum'] = self.max_executor_num
 
         if self.pool_name is not None:
             result['PoolName'] = self.pool_name
@@ -134,18 +142,27 @@ class ListPoolsResponseBodyPoolList(DaraModel):
         if self.priority is not None:
             result['Priority'] = self.priority
 
+        if self.scheduling_policy_id is not None:
+            result['SchedulingPolicyId'] = self.scheduling_policy_id
+
         if self.status is not None:
             result['Status'] = self.status
+
+        if self.update_time is not None:
+            result['UpdateTime'] = self.update_time
 
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('CreateTime') is not None:
+            self.create_time = m.get('CreateTime')
+
         if m.get('IsDefault') is not None:
             self.is_default = m.get('IsDefault')
 
-        if m.get('MaxExectorNum') is not None:
-            self.max_exector_num = m.get('MaxExectorNum')
+        if m.get('MaxExecutorNum') is not None:
+            self.max_executor_num = m.get('MaxExecutorNum')
 
         if m.get('PoolName') is not None:
             self.pool_name = m.get('PoolName')
@@ -153,8 +170,14 @@ class ListPoolsResponseBodyPoolList(DaraModel):
         if m.get('Priority') is not None:
             self.priority = m.get('Priority')
 
+        if m.get('SchedulingPolicyId') is not None:
+            self.scheduling_policy_id = m.get('SchedulingPolicyId')
+
         if m.get('Status') is not None:
             self.status = m.get('Status')
+
+        if m.get('UpdateTime') is not None:
+            self.update_time = m.get('UpdateTime')
 
         return self